requirements.txt
joblib==1.3.2
matplotlib==3.7.3
numpy==1.24.4
pandas==2.0.3
Pillow==10.0.0
rasterio==1.3.7
requests==2.31.0
scikit-learn==1.3.0
scipy==1.10.1
sentinelhub==3.9.1
seaborn==0.12.2
tqdm==4.66.1
xgboost==1.7.6
lxml==4.9.3
openpyxl==3.1.2
azure-storage-blob==12.18.3
python-dotenv==1.0.0
simplekml==1.3.6
fastkml==0.12
torch==2.0.0
earthengine-api
geopandas
